requirements.txt
streamlit
openai
qdrant-client
langchain
langchain_openai
python-dotenv
pydantic
pandas
python-docx
PyPDF2
nltk
tabulate
requests
pydantic_ai
langgraph
openpyxl
numpy
streamlit-webrtc
pydub
streamlit-mic-recorder